stateNode changed to initialNode
